+++ encore_api_cli/commands/upload.py
@@ -14,7 +14,7 @@ def cli() -> None:  # noqa: D103
     pass
 
 
-@cli.command()
+@cli.command(short_help="Upload the local movie or image file to the cloud storage.")
 @click.argument("path", type=click.Path(exists=True))
 @common_options
 @pass_state